server restructuring
