+++ src/routers/loginRouter.js
@@ -2,7 +2,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const jsonParser = bodyParser.json();
 const router = express.Router();
-const {User} = require('./models');
+const {User} = require('../models');
 
 router.get('/', (req, res) => {
   res.sendFile(__dirname + '/public/login.html');
